requirements.txt
fastapi
uvicorn[gunicorn]
sqlalchemy
psycopg2-binary
python-dotenv